+++ src/app/sections/auth/login/components/LoginForm.tsx
@@ -48,47 +48,51 @@ const LoginForm: React.FC = () => {
 };
 
   return (
-    <form onSubmit={handleLogin} className="w-full max-w-xl bg-white border-gray-500 items-center shadow-xl rounded-2xl p-8 flex flex-col justify-between min-h-[300px]">
-      <h2 className="text-xl font-semibold text-gray-700 text-center p-6">
-        Bienvenido a SkilLogin Academy
-      </h2>
+    <div className="flex flex-col h-full w-full max-w-xl mx-auto px-4">
+      <div className="bg-white border-blue-300 justify-center shadow-lg border-2 rounded-2xl w-full overflow-hidden">
+        <h3 className="text-xl font-semibold text-gray-700 text-center p-6">Inicio de sesión</h3>
 
-      {error && <p className="text-red-600 text-sm text-center">{error}</p>}
+        {/* Form with login data*/}
+        <form onSubmit={handleLogin} className="w-full max-w-xl bg-white border-gray-500 items-center shadow-xl rounded-2xl p-8 flex flex-col justify-between min-h-[300px]">
+         
+          {error && <p className="text-red-600 text-sm text-center">{error}</p>}
 
-      <div>
-        <label className="block text-sm mb-2 font-medium text-gray-700">Email</label>
-        <input
-          type="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          className="px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-          required
-        />
+          <div>
+            <label className="block text-sm mb-2 font-medium text-gray-700">Email</label>
+            <input
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              className="px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">Password</label>
+            <input
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              className="w-full p-2 border rounded"
+              required
+            />
+          </div>
+
+          <button
+            type="submit"
+            className="w-full bg-blue-600 text-white py-2 rounded-md hover:bg-blue-700 transition"
+          >
+            Login
+          </button>
+
+          <p className="text-center text-sm text-gray-600">
+            ¿No tienes una cuenta?{' '}
+            <Link  href="/?view=register" className="text-blue-600 underline">Crear una cuenta </Link>
+          </p>
+        </form>
       </div>
-
-      <div>
-        <label className="block text-sm font-medium text-gray-700">Password</label>
-        <input
-          type="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          className="w-full p-2 border rounded"
-          required
-        />
-      </div>
-
-      <button
-        type="submit"
-        className="w-full bg-blue-600 text-white py-2 rounded-md hover:bg-blue-700 transition"
-      >
-        Login
-      </button>
-
-      <p className="text-center text-sm text-gray-600">
-        ¿No tienes una cuenta?{' '}
-        <Link  href="/?view=register" className="text-blue-600 underline">Crear una cuenta </Link>
-      </p>
-    </form>
+    </div>
   );
 };
